requirements.txt
numpy
ultralytics
opencv-python
Pillow